# speedups in pT3 