- "massage" the implementation in `createPixelTripletsInGPUFromMapv2` and called methods to reduce the number of registers: 
    - the number of used registers as in ptxas info reduced from 138 to 124. 
        - At some point rather similar code had 118 registers, but I lost track of that version; it was not significantly faster 
    - pT3 time changed from 5.5 to 5.2 ms
    - `createPixelTripletsInGPUFromMapv2` reduced from 1.25 ms to 1.02 ms
- remove (pLS X connectedModules) remapping/flattening done on CPU and simply use another dimension in the grid to do it on the fly; this saves the cost of mem copy of about 8MB of data
    - this provides the largest improvement in timing: pT3 time (on phi3) goes from 5.2 ms to 1.7 ms (compared from 8b578a3b385 to dacd609f8)

The validation was limited: looking at just the number of produced pT3s and other objects