@slava77 2022-05-12
time for 20 evts on phi3: (ef72c6c is old, dacd609 is new)
```
   Evt    Hits       MD       LS      T3       T5       pLS       pT5      pT3      TC       TCE      Event      Short             Rate
   old      3.0      4.0      3.5      3.5      3.7      5.0      6.4      5.5      6.9      5.8      47.1      39.1+/-  6.6      52.2   explicit_cache[s=1]
   new      2.8      3.8      3.4      3.4      3.7      5.0      6.5      1.7      6.3      5.6      42.2      34.5+/-  5.8      45.8   explicit_cache[s=1]
```