@tresreid 2022-05-12
validation looks good: https://www.classe.cornell.edu/~mgr85/www/SegmentLinking_keep/PR159/eff_plots__GPU_explicit_cache_dacd609_PU200/#1.5